% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2disease.R
\name{gene2disease}
\alias{gene2disease}
\title{gene2disease}
\usage{
gene2disease(gene)
}
\arguments{
\item{gene}{gene name}
}
\description{
A function that takes the prepared disease target genetic association and match them with gene names from feline GWAS
}
