% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2diseaseMan.R
\name{gene2diseaseMan}
\alias{gene2diseaseMan}
\title{gene2diseaseMan}
\usage{
gene2diseaseMan(gene)
}
\arguments{
\item{gene}{gene name}
}
\description{
A function that takes the prepared disease target genetic association and match them with gene names from feline GWAS
}
