% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticAssocPrep.R
\name{geneticAssocPrep}
\alias{geneticAssocPrep}
\title{geneticAssocPrep}
\usage{
geneticAssocPrep(assocOT, diseaseOT, humanGTF, ResultsPath = ".")
}
\arguments{
\item{assocOT}{evidence by data source file full path (see external documentation for data download from OpenTargets).}

\item{diseaseOT}{Disease data full path(see external documentation for data download from OpenTargets)}

\item{humanGTF}{human GTF file from GENCODE}

\item{ResultsPath}{default is the working directory but you can provide a path of your own ensuring it ends with a /}
}
\description{
Preparation of OpenTarget data evidence by data source file for gwasFollowupMan function
}
