% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasFollowupFull.R
\name{gwasFollowupFull}
\alias{gwasFollowupFull}
\title{gwasFollowupFull}
\usage{
gwasFollowupFull(
  sumStats,
  felGTF,
  pval = 5e-08,
  ResultsPath = ".",
  phenomePath = NULL,
  twasPath = NULL
)
}
\arguments{
\item{sumStats}{GWAS summary statistics file. It assumed a ps, and chr columns.}

\item{felGTF}{GTF file}

\item{pval}{default is 5*10-8 and you could change but make sure to use standard form}

\item{ResultsPath}{default is the working directory but you can provide a path of your own ensuring it ends with a /}

\item{phenomePath}{path to phenomeXcan data if available default NULL}

\item{twasPath}{path to twas data if available default NULL}
}
\description{
main package function that will take GWAS sumstats, a gtf file, and a pvalue of interest to returns information on the genes of interest
}
