% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasFollowupMan.R
\name{TargetChaseManual}
\alias{TargetChaseManual}
\title{TargetChaseManual}
\usage{
TargetChaseManual(
  sumStats,
  gtf,
  species = "cat",
  pval = 5e-08,
  ResultsPath = ".",
  impc = NULL,
  assocOT,
  l2gOT,
  zoo_dir = NULL
)
}
\arguments{
\item{sumStats}{GWAS summary statistics file. Must have columns: ps, chr, p_wald.}

\item{gtf}{GTF file for the species}

\item{species}{Species of the GWAS data: "human", "cat", or "dog". Default is "cat".}

\item{pval}{P-value threshold. Default is 5e-8.}

\item{ResultsPath}{Directory to write output files. Created if it doesn't exist.}

\item{impc}{IMPC data file path (from IMPCprep or downloadData)}

\item{assocOT}{OpenTargets genetic association file path (from geneticAssocPrep or downloadData)}

\item{l2gOT}{OpenTargets locus-to-gene file path (from l2gPrep or downloadData)}

\item{zoo_dir}{Directory containing Zoonomia orthology files. If NULL, uses package extdata.}
}
\value{
Writes g2d_results.txt and l2g_results.txt to ResultsPath
}
\description{
Extension function that takes GWAS sumstats, a GTF file, and a p-value threshold,
along with locally prepared OpenTargets and IMPC data files. Use this function when
you have pre-downloaded and prepared genetic association and l2g data from OpenTargets.
}
