% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2gPrep.R
\name{l2gPrep}
\alias{l2gPrep}
\title{l2gPrep}
\usage{
l2gPrep(l2gOT, studyOT, humanGTF, ResultsPath = ".")
}
\arguments{
\item{l2gOT}{l2g Data full path (see external documentation for data download from OpenTargets).}

\item{studyOT}{study index data full path(see external documentation for data download from OpenTargets)}

\item{humanGTF}{human GTF file from GENCODE}

\item{ResultsPath}{default is the working directory but you can provide a path of your own ensuring it ends with a /}
}
\description{
Preparation of OpenTarget l2g gwasFollowupMan function
}
