% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus2gene.R
\name{locus2gene}
\alias{locus2gene}
\title{locus2gene}
\usage{
locus2gene(gene)
}
\arguments{
\item{gene}{gene name}
}
\description{
A function that takes the prepared l2g data from opentargets match them with gene names from feline GWAS
}
