% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus2geneMan.R
\name{locus2geneMan}
\alias{locus2geneMan}
\title{locus2geneMan}
\usage{
locus2geneMan(gene)
}
\arguments{
\item{gene}{gene name}
}
\description{
A function that takes the prepared l2g data from opentargets match them with gene names from feline GWAS
}
