% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomePrep.R
\name{phenomePrep}
\alias{phenomePrep}
\title{phenomePrep}
\usage{
phenomePrep(phenome, gtf, ResultsPath = ".")
}
\arguments{
\item{phenome}{path to bulk phenomeXcan data}

\item{gtf}{path to human gtf file (use recommended)}

\item{ResultsPath}{default is the working directory but you can provide a path of your own ensuring it ends with a /}
}
\description{
Prepare PhenomeXcan data if present
}
