% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasFollowupMan.R
\name{gwasFollowupMan}
\alias{gwasFollowupMan}
\title{gwasFollowupMan}
\usage{
gwasFollowupMan(
  sumStats,
  felGTF,
  pval = 5e-08,
  ResultsPath = ".",
  impc = NULL,
  assocOT,
  l2gOT
)
}
\arguments{
\item{sumStats}{GWAS summary statistics file. It assumed a ps, and chr columns.}

\item{felGTF}{GTF file}

\item{pval}{default is 5*10-8 and you could change but make sure to use standard form}

\item{ResultsPath}{default is the working directory but you can provide a path of your own ensuring it ends with a /}

\item{impc}{impc data}

\item{assocOT}{results from geneticAssocPrep function}

\item{l2gOT}{results from l2gPrep function}
}
\description{
main package extensionfunction that will take GWAS sumstats, a gtf file, and a pvalue of interest to returns information on the genes of interest but also the genetic
association file and l2g file data prepped with l2gPrep and geneticAssocPrep functions in this package. Only use this function when you have updated genetic and l2g data
from opentargets
}
